src/main/webapp/app/entities/security-group/route/security-group-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ISecurityGroup, SecurityGroup } from '../security-group.model';
import { SecurityGroupService } from '../service/security-group.service';

import { SecurityGroupRoutingResolveService } from './security-group-routing-resolve.service';

describe('SecurityGroup routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SecurityGroupRoutingResolveService;
  let service: SecurityGroupService;
  let resultSecurityGroup: ISecurityGroup | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(SecurityGroupRoutingResolveService);
    service = TestBed.inject(SecurityGroupService);
    resultSecurityGroup = undefined;
  });

  describe('resolve', () => {
    it('should return ISecurityGroup returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSecurityGroup = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSecurityGroup).toEqual({ id: 123 });
    });

    it('should return new ISecurityGroup if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSecurityGroup = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSecurityGroup).toEqual(new SecurityGroup());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as SecurityGroup })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSecurityGroup = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSecurityGroup).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
